Rename special rendered attributes
